refactor: move TestConfig's package in integration package
