+++ src/test/java/com/example/spinlog/integration/test_list.md
@@ -1,0 +1,15 @@
+### 캐시 관련 기능
+
+1. Article 업데이트 시, 성별 통계 캐시 업데이트
+2. User 업데이트 시, 성별 통계 캐시 업데이트
+3. 매일 04:00에 성별 통계 캐시 업데이트 - have to lock(PERIOD_CRITERIA)
+4. 매일 05:00에 캐시 검증
+5. 성별 통계 API로 캐시 조회할 때 에러 발생 시, fallback 처리(레포지토리 조회, 캐시 업데이트)
+
++ lock 테스트  
++ 0~4시 데이터 정합성 테스트
+
+---
+
+### 통합 테스트 목록
+
